@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <RootNamespace>Faemiyah.BtDamageResolver.Actors</RootNamespace>
     <AssemblyName>Faemiyah.BtDamageResolver.Actors</AssemblyName>
   </PropertyGroup>
@@ -13,7 +13,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Orleans.Runtime" Version="7.2.1" />
+    <PackageReference Include="Microsoft.Orleans.Runtime" Version="7.2.3" />
   </ItemGroup>
 
   <ItemGroup>
