@@ -1,14 +1,14 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <RootNamespace>Faemiyah.BtDamageResolver.Services</RootNamespace>
     <AssemblyName>Faemiyah.BtDamageResolver.Services</AssemblyName>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Orleans.Runtime" Version="7.2.1" />
-    <PackageReference Include="Npgsql" Version="7.0.6" />
+    <PackageReference Include="Microsoft.Orleans.Runtime" Version="7.2.3" />
+    <PackageReference Include="Npgsql" Version="8.0.0" />
   </ItemGroup>
 
   <ItemGroup>
