@@ -2,19 +2,19 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <RootNamespace>Faemiyah.BtDamageResolver.Silo</RootNamespace>
     <ServerGarbageCollection>true</ServerGarbageCollection>
     <ConcurrentGarbageCollection>true</ConcurrentGarbageCollection>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="7.0.1" />
-    <PackageReference Include="Microsoft.Orleans.Clustering.AdoNet" Version="7.2.1" />
-    <PackageReference Include="Microsoft.Orleans.Persistence.AdoNet" Version="7.2.1" />
-    <PackageReference Include="Microsoft.Orleans.Sdk" Version="7.2.1" />
-    <PackageReference Include="Microsoft.Orleans.Serialization.NewtonsoftJson" Version="7.2.1" />
-    <PackageReference Include="Microsoft.Orleans.Server" Version="7.2.1" />
+    <PackageReference Include="Microsoft.Extensions.Hosting" Version="8.0.0" />
+    <PackageReference Include="Microsoft.Orleans.Clustering.AdoNet" Version="7.2.3" />
+    <PackageReference Include="Microsoft.Orleans.Persistence.AdoNet" Version="7.2.3" />
+    <PackageReference Include="Microsoft.Orleans.Sdk" Version="7.2.3" />
+    <PackageReference Include="Microsoft.Orleans.Serialization.SystemTextJson" Version="7.2.3" />
+    <PackageReference Include="Microsoft.Orleans.Server" Version="7.2.3" />
   </ItemGroup>
 
   <ItemGroup>
