@@ -1,10 +1,10 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net7.0</TargetFramework>
-    <RootNamespace>Faemiyah.BtDamageResolver.Tests</RootNamespace>
-    <AssemblyName>CompressionTesterApp</AssemblyName>
+    <TargetFramework>net8.0</TargetFramework>
+    <RootNamespace>Faemiyah.BtDamageResolver.CompressionTesterApp</RootNamespace>
+    <AssemblyName>Faemiyah.BtDamageResolver.CompressionTesterApp</AssemblyName>
   </PropertyGroup>
 
   <ItemGroup>
