@@ -2,18 +2,18 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <RootNamespace>Faemiyah.BtDamageResolver.Tools.DataImporter</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="CommandLineParser" Version="2.9.1" />
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="7.0.1" />
-    <PackageReference Include="Microsoft.Orleans.Client" Version="7.2.1" />
-    <PackageReference Include="Microsoft.Orleans.Clustering.AdoNet" Version="7.2.1" />
-    <PackageReference Include="Microsoft.Orleans.Sdk" Version="7.2.1" />
-    <PackageReference Include="Microsoft.Orleans.Serialization.NewtonsoftJson" Version="7.2.1" />
-    <PackageReference Include="Npgsql" Version="7.0.6" />
+    <PackageReference Include="Microsoft.Extensions.Hosting" Version="8.0.0" />
+    <PackageReference Include="Microsoft.Orleans.Client" Version="7.2.3" />
+    <PackageReference Include="Microsoft.Orleans.Clustering.AdoNet" Version="7.2.3" />
+    <PackageReference Include="Microsoft.Orleans.Sdk" Version="7.2.3" />
+    <PackageReference Include="Microsoft.Orleans.Serialization.SystemTextJson" Version="7.2.3" />
+    <PackageReference Include="Npgsql" Version="8.0.0" />
   </ItemGroup>
 
   <ItemGroup>
