@@ -1,5 +1,6 @@
 ﻿using System.Text;
-using Newtonsoft.Json;
+using System.Text.Json;
+using Microsoft.Extensions.Options;
 using SevenZip.Compression.LZMA;
 
 namespace Faemiyah.BtDamageResolver.Api.ClientInterface.Compression;
@@ -7,15 +8,26 @@ namespace Faemiyah.BtDamageResolver.Api.ClientInterface.Compression;
 /// <summary>
 /// Static easy-to-use compression helping methods for compressing complex types through JSON serialization.
 /// </summary>
-public static class DataHelper
+public class DataHelper
 {
+    private readonly JsonSerializerOptions _jsonSerializerOptions;
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="DataHelper"/> class.
+    /// </summary>
+    /// <param name="jsonSerializerOptions">The JSON serializer options.</param>
+    public DataHelper(IOptions<JsonSerializerOptions> jsonSerializerOptions)
+    {
+        _jsonSerializerOptions = jsonSerializerOptions.Value;
+    }
+
     /// <summary>
     /// Serialize an entity and compress the byte array.
     /// </summary>
     /// <typeparam name="TType">The type to compress.</typeparam>
     /// <param name="input">The byte array to compress.</param>
     /// <returns>The compressed byte array.</returns>
-    public static byte[] Pack<TType>(TType input)
+    public byte[] Pack<TType>(TType input)
         where TType : class
     {
         return CompressionHelper.Compress(Serialize(input));
@@ -24,10 +36,10 @@ public static class DataHelper
     /// <summary>
     /// Decompress a byte array and deserialize entity.
     /// </summary>
-    /// <typeparam name="TType">The type to uncompress.</typeparam>
     /// <param name="input">The byte array to decompress.</param>
+    /// <typeparam name="TType">The compressed type.</typeparam>
     /// <returns>The decompressed byte array.</returns>
-    public static TType Unpack<TType>(byte[] input)
+    public TType Unpack<TType>(byte[] input)
         where TType : class
     {
         return Deserialize<TType>(CompressionHelper.Decompress(input));
@@ -39,10 +51,10 @@ public static class DataHelper
     /// <typeparam name="TType">The type of the object.</typeparam>
     /// <param name="input">The byte array.</param>
     /// <returns>The deserialized object.</returns>
-    private static TType Deserialize<TType>(byte[] input)
+    private TType Deserialize<TType>(byte[] input)
         where TType : class
     {
-        return JsonConvert.DeserializeObject<TType>(Encoding.UTF8.GetString(input));
+        return JsonSerializer.Deserialize<TType>(Encoding.UTF8.GetString(input), _jsonSerializerOptions);
     }
 
     /// <summary>
@@ -51,9 +63,9 @@ public static class DataHelper
     /// <typeparam name="TType">The type of the object.</typeparam>
     /// <param name="input">The object.</param>
     /// <returns>The serialized byte array.</returns>
-    private static byte[] Serialize<TType>(TType input)
+    private byte[] Serialize<TType>(TType input)
         where TType : class
     {
-        return Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(input));
+        return Encoding.UTF8.GetBytes(JsonSerializer.Serialize(input, _jsonSerializerOptions));
     }
 }
