@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Net;
+using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
 using Faemiyah.BtDamageResolver.Actors.Cryptography;
@@ -7,6 +8,7 @@ using Faemiyah.BtDamageResolver.Actors.Logic;
 using Faemiyah.BtDamageResolver.Actors.Logic.ExpressionSolver;
 using Faemiyah.BtDamageResolver.Actors.Logic.Interfaces;
 using Faemiyah.BtDamageResolver.Api;
+using Faemiyah.BtDamageResolver.Api.ClientInterface.Compression;
 using Faemiyah.BtDamageResolver.Api.ClientInterface.Repositories;
 using Faemiyah.BtDamageResolver.Api.Entities.Interfaces;
 using Faemiyah.BtDamageResolver.Api.Entities.RepositoryEntities;
@@ -20,7 +22,6 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using Newtonsoft.Json;
 using Orleans;
 using Orleans.Configuration;
 using Orleans.Hosting;
@@ -118,7 +119,7 @@ public static class Program
                 siloBuilder
                     .Services.AddSerializer(serializerBuilder =>
                     {
-                        serializerBuilder.AddNewtonsoftJsonSerializer(isSupported: type => type.Namespace != null && type.Namespace.StartsWith("Faemiyah.BtDamageResolver"));
+                        serializerBuilder.AddJsonSerializer(isSupported: type => type.Namespace != null && type.Namespace.StartsWith("Faemiyah.BtDamageResolver"));
                     });
                 siloBuilder
                     .Configure<ClusterOptions>(options =>
@@ -145,12 +146,6 @@ public static class Program
                         options.ResponseTimeout = TimeSpan.FromSeconds(15);
                         options.ResponseTimeoutWithDebugger = TimeSpan.FromMinutes(15);
                     })
-                    .Configure<JsonSerializerSettings>(options =>
-                    {
-                        options.MissingMemberHandling = MissingMemberHandling.Error;
-                        options.NullValueHandling = NullValueHandling.Ignore;
-                        options.Converters.Add(new Newtonsoft.Json.Converters.StringEnumConverter());
-                    })
                     .UseAdoNetClustering(options =>
                     {
                         options.Invariant = clusterOptions?.Invariant;
@@ -172,6 +167,7 @@ public static class Program
                     .AddGrainService<LoggingService>()
                     .ConfigureServices(services =>
                     {
+                        services.ConfigureJsonSerializerOptions();
                         services.Configure<CommunicationOptions>(configuration.GetSection(Settings.CommunicationOptionsBlockName));
                         services.Configure<FaemiyahClusterOptions>(configuration.GetSection(Settings.ClusterOptionsBlockName));
                         services.Configure<FaemiyahLoggingOptions>(configuration.GetSection(Settings.LoggingOptionsBlockName));
@@ -201,6 +197,7 @@ public static class Program
                         services.AddSingleton<CachedEntityRepository<PaperDoll, string>, CachedEntityRepository<PaperDoll, string>>();
                         services.AddSingleton<CachedEntityRepository<Unit, string>, CachedEntityRepository<Unit, string>>();
                         services.AddSingleton<CachedEntityRepository<Weapon, string>, CachedEntityRepository<Weapon, string>>();
+                        services.AddSingleton<DataHelper>();
                     });
                     /*.AddStartupTask((serviceProvider, _) =>
                     {
@@ -231,7 +228,7 @@ public static class Program
     {
         var options = serviceProvider.GetService<IOptions<CommunicationOptions>>();
         return options != null
-            ? new RedisEntityRepository<TType>(serviceProvider.GetService<ILogger<RedisEntityRepository<TType>>>(), serviceProvider.GetService<IOptions<JsonSerializerSettings>>(), options.Value.ConnectionString)
+            ? new RedisEntityRepository<TType>(serviceProvider.GetService<ILogger<RedisEntityRepository<TType>>>(), serviceProvider.GetService<IOptions<JsonSerializerOptions>>(), options.Value.ConnectionString)
             : throw new InvalidOperationException($"Unable to resolve options class providing connection string for entity repository of type {typeof(TType)}.");
     }
 
