@@ -20,7 +20,8 @@ public interface IServerToClientCommunicator
     /// <remarks>
     /// Envelope type is a processing hint for the recipient.
     /// </remarks>
-    public void Send<TType>(string clientName, string envelopeType, TType data);
+    public void Send<TType>(string clientName, string envelopeType, TType data)
+        where TType : class;
 
     /// <summary>
     /// Sends data to all clients.
@@ -31,7 +32,8 @@ public interface IServerToClientCommunicator
     /// <remarks>
     /// Envelope type is a processing hint for the recipient.
     /// </remarks>
-    public void SendToAll<TType>(string envelopeType, TType data);
+    public void SendToAll<TType>(string envelopeType, TType data)
+        where TType : class;
 
     /// <summary>
     /// Sends data to many clients.
@@ -43,7 +45,8 @@ public interface IServerToClientCommunicator
     /// <remarks>
     /// Envelope type is a processing hint for the recipient.
     /// </remarks>
-    public void SendToMany<TType>(List<string> clientNames, string envelopeType, TType data);
+    public void SendToMany<TType>(List<string> clientNames, string envelopeType, TType data)
+        where TType : class;
 
     /// <summary>
     /// Handle an incoming <see cref="ConnectRequest"/>.
