@@ -1,5 +1,4 @@
 ﻿using System;
-using static Faemiyah.BtDamageResolver.Api.ClientInterface.Compression.DataHelper;
 
 namespace Faemiyah.BtDamageResolver.Api.ClientInterface;
 
@@ -28,9 +27,9 @@ public class Envelope
     /// <remarks>
     /// The data type hint is a hint for the recipient on how to process the data.
     /// </remarks>
-    public Envelope(string dataTypeHint, object data)
+    public Envelope(string dataTypeHint, byte[] data)
     {
-        Data = Pack(data);
+        Data = data;
         CorrelationId = Guid.NewGuid();
         TimeStamp = DateTime.UtcNow;
         Type = dataTypeHint;
